(PC-31176) refactor(booking): extract reaction icon logic in external hook
